Remaking request method. Now we use Alamofire with new model data which is class. Make some fix for pull request. And FlickrLayouts actually works but not as we need...
